Remove toggle interval functionality; Improvements
No longer required to change sample rate. Diagrams and code outline updated to remove this functionality.

In the code there was a clear_log() fn that cleared the eeprom. It has been renamed clear_mem() so that
a new fn clear_log() can be used to clear the console.

Use case diagram has been improved.